QCachedPainter: use the render hints of the painter we got
Amends 1ad9330a9c56a59d407f5644fe547ba69c1a205b. If the painter that we
get from the style (or the widget) has render hints set, then we want to
use those when drawing the cached pixmap as well.

This assumes that QPixmap as a paint device supports all the required
engine features, which it does as it's using the software rasterizer.

This fixes baseline test mismatches for QDial and QSlider, as those want
to be rendered using an anti-aliased painter.

Change-Id: Iaf9ec1c3a4f940f31fb34badc436e499e19f6698
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>