Tweak QInputDevice logging
The logging category now silences QtWarningMsg and below by
default, so we can be more granular in our logging.

Use info logging category for registering/unregistering, to allow
filtering out the detailed debug logging while still seeing when
devices are registered/unregistered.

Log the number of buttons for pointing devices, if there are any.

Log the device for wheel events.

Pick-to: 6.8
Change-Id: Ie046fee1659fc3967caf7eafbaf46f3ed22be5bb
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>