JNI: Optimize string handling
A lot of code passes strings through JNI calls, and we always ended up
constructing a temporary QJniObject just to convert it to a QString, or
to get the jstring out of it. This is expensive as QJniObject allocates
a private object and maintains a reference count.

For code paths that have a jstring, or only want a jstring, and already
have a JNIEnv pointer, we can shortcut this and use JNI APIs directly
rather than going through the QJniObject detour. This was already done
when accessing elements of a QJniArray<QString>.

Add internal helper methods to the QtJniTypes::Details namespace that
convert between jstring and QString directly. Use that where we can,
including the QJniObject::to/fromString implementations.

Since the new helper for converting from a QString already returns a
local reference, split the LocalFrame::newLocalRef into a helper that
pushes the local frame, and use that to explicitly do so in the case
of QString conversion. This also removes the last use of the
QJniObject overload of LocalFrame::newLocalRef, so remove that.

Change-Id: Ib3a261efd0dcca86a1a99c2a054d50e932c3fe71
Reviewed-by: Soheil Armin <soheil.armin@qt.io>
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>