JNI: Optimize QJniObject string conversions
If we know that the QJniObject references a jstring, then we can fast-
track the conversion to a QString. Any QJniObject that was created via
fromString references a jstring; and any object with the class name
"java/lang/String", or where the jclass is identical to the one for
jstring, holds a jstring. Implement a helper that makes fast checks (in
the worst case, a string comparison).

If we know that we have a jstring, then we don't have to call the
toString() Java method on the object to serialize it to a jstring,
we can directly operate on the referenced jobject as a jstring.

Avoid duplicating the "java/lang/String" string and instead use the
compile-time class name we get for jstring from QtJniTypes.

Change-Id: I46895b6c4fab67a9ff4f6ccfadd9daf5b4911fab
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>