Fix assertion failure in QLocale::uiLanguages()
If an entry in the list from the system query exercises a legacy tag,
the QLocaleId it gets converted to has a name() that's different from
it. The check that this name matches the list item that it's derived
from was consequently failing. Have that check fall back to actually
doing the round-trip via parsing to compare the name, if the (still
cheaper, so retained) naive check fails.

Fixes: QTBUG-131127
Change-Id: Ifa004deff880109031ff56be00bebff6f5106e62
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>