Fix flaky QtAbstractItemModel setData test case
The Java callback function onDataChanged() gets called
from Android UI thread. QTRY_COMPARE on the result of
a JNI call does not make sense because it actually does
not retry to reinvoke the JNI call to read the related
counter.

This change puts the JNI call to QTRY_COMPARE to make sure
the Android thread have to to process the Runnable.

Fixes: QTBUG-131253
Change-Id: Icdff8f8f37b33968f5a3f6e545ce753907305e74
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>