wasm: streamline pointer event handling
We know the type of pointer event based on the callback
(down, move, up, etc), which means the code which determines
the type based on the type string can be removed.

This saves one string comparison per event, and also
allows removing the runtime error case where we are unable
to determine which type of pointer event we have.

Remove PointerEvent::fromWeb(), and construct the pointer
event directly in the event handler instead.

Change-Id: I4fe566bf076dddd5bf39217f1d6671ba07e25912
Reviewed-by: Morten Johan Sørvig <morten.sorvig@qt.io>