QFontIconEngine: render named glyphs as icons, if possible
If the QIcon::themeName matches an installed font, and if the
name of the icon matches a named glyph in the font, then render
that glyph as a painter path.

Overrides of QFontIconEngine::text() take priority.

Amend the manual test to allow specifying an icon theme on the
command line, and render the named glyph also as text, as
some icon fonts will define ligatures that turn the string into
the corresponding icon.

Task-number: QTBUG-102346
Change-Id: I788c6274322359955cbfe58175a2999a57cfce95
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>