CMake: Move separate debug info handling to an executable finalizer
Currently the separate debug info handling is done inside the call of
qt_internal_add_executable. But there might be extra properties set
after the executable is created, which we might want to consider when
handling the separate debug info, like whether the executable is a
macOS bundle.

Introduce a new qt_internal_finalize_executable finalizer. Move the
separate debug info handler to be called in this finalizer.

To be consistent run the separate debug info handler in a finalizer
for qt tools as well.

Pick-to: 6.8
Change-Id: I46412249aaab099628a50b11efff541f5719aff5
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>