Fix blacklisting of QNetworkInterface::localAddress data tags
When creating blacklists for parametric tests, I've misunderstand the
role of comma in parametric tests descriptor in BLACKLIST, thinking that
this is a separator when in fact it was just a part of parameter name.
This leads to tests not being actually blacklisted for VxWorks.

Fix this by blacklisting each data tag for test function separately.

Task-number: QTBUG-115777
Task-number: QTBUG-130070
Change-Id: Id282c361d965e24341cd656aec6b1514e2c1f84f
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>