Add QEXPECT_FAIL to tst_QFile::openDirectory() for VxWorks
Opening existing directory with write mode always passes on VxWorks, the
reason is to be found. This causes the failure of
tst_QFile::openDirectory test function.

We need to run tests on COIN for VxWorks as significant, and fail in
QFile::openDirectory test functions prevents it, to reduce growing
number of regressions and allow iterative introduction of VxWorks as a
supported platform.

Add QEXPECT_FAIL macro to tst_QFile::openDirectory() test function for
VxWorks before verifying if opening of directory with ReadWrite
permissions works.

Task-number: QTBUG-115777
Task-number: QTBUG-130074
Change-Id: Iba8bd5ed143d43ddd665b213cc850af503e476e2
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>