Expect fails in QApplication test functions for VxWorks
"libraryPaths" and "libraryPaths_qt_plugin_path" test functions fail on
VxWorks, due to missing path to "/usr/local/Qt-6.9.0/plugins" which test
expects.

The exact reason of this failure is to be found.

We need to run tests on COIN for VxWorks as significant, and this test
prevents it, to reduce growing number of regressions and allow
iterative introduction of VxWorks as a supported platform.

Add QEXPECT_FAIL in library-path-related test function of
tst_QApplication for VxWorks.

Task-number: QTBUG-115777
Task-number: QTBUG-130736
Change-Id: I08bb8b3f4a776303f61aebd931e5f3ddf3f1f39e
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>