QFontEngine: add API to get the name of a glyph
Some fonts name their glyphs in the "post" or "cff" tables, and both
CoreText and FreeType provide high-level APIs to access those values.
Also, harfbuzz has an API for OpenType fonts, which we can use on
Windows (where we would have to load the font into FreeType, as
neither GDI nor DirectWrite provide APIs to access this data), and as
default implementation that we can fall back to if the platform APIs
don't find anything.

We need to set the harfbuzz font up with the OpenType callbacks
explicitly, so cannot just use what we get from harfbuzzFont.

With this low-level API we can now make name-based lookups of glyphs,
and eventually render those in a QIconEngine implementation.

Task-number: QTBUG-102346
Change-Id: I68cc19814fc45d63a88e063b719b46f6aa6100bc
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>