rhi: Make DXGI HDR query usable with Vulkan
Also moves the relevant code into one place, that is usable internally
from other places in the future. (the interface might change still)

Most importantly, this now makes the Vulkan backend to report the same
information as the D3D11/12 ones, as long as the Vulkan implementation
reports the deviceLUID in the extended physical device properties.

Change-Id: I6aa8adaf40bbc3199959019d323910abd27ec473
Reviewed-by: Kristoffer Skau <kristoffer.skau@qt.io>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>