QLabel: clearContents() when setting a new QPixmap
When a label is displaying a QPixmap after it showed text, the pixmap
was not shown as the label still was in 'text-mode'. Add the missing
clearContents() before setting a new QPixmap. This was removed by
accident with 81e74cfa2610fb2aa90b2352d498762d46d0ddd5.

This amends 81e74cfa2610fb2aa90b2352d498762d46d0ddd5.

Task-number: QTBUG-122403
Change-Id: I024df6a17458f98f715eba13caaffb3213d0e9fa
Reviewed-by: Nodir Temirkhodjaev <nodir.temir@gmail.com>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>