QFusionStyle/CC_Slider: consistently turn anti-aliasing on
Amends e5f9738385d8fc49492bc759472a571e241a5501, which removed the
turning on of anti-aliasing, under the assumption that QCachedPainter
always turns on anti-aliasing (which was initially the case, but then
reverted in 1ad9330a9c56a59d407f5644fe547ba69c1a205b). On the other
hand, the later QCachedPainter instances still turned on anti-aliasing
explicitly.

Now we do so again consistently, but setting it on the painter we get;
QCachedPainter will then use the same render hints.

As a drive-by, constify the variables with the old painter states
consistently.

Change-Id: I2ed31ae42c77f07d8496dcb16b767def29203973
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>