visionOS: Adapt the rhi setViewport function to work with rate maps
When shading rate maps are used we shouldn't use the target's pixel
size to set the viewport, as the target's pixel size is likely to
different then the "output" size. Instead we get the viewport size from
Metal's screenSize function. The rendered content should then be
correctly scaled to the output size by the rate map.

Pick-to: 6.8
Task-number: QTBUG-129733
Change-Id: I1c53ded0a700eb860cd3bf08fc12a0404c2e6ffa
Reviewed-by: Andy Nichols <andy.nichols@qt.io>