CMake: Fix CMP0174 warnings in Qt6CTestMacros.cmake
Invalid build dir values were passed to the WORKING_DIRECTORY option
to the qt test apis.

We shouldn't be passing the option at all, it makes no difference for
the case of calling ctest --build-and-test, because we already specify
the build dir in the command line of ctest.

Amends bc5b7e218f7259f440449e69c42518467bc5907b

Pick-to: 6.8
Change-Id: I0fe5b20a4beaefbd129849699c873c6a558d8996
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>