wasm: Stop configuration if Emscripten versions don't match
Emscripten compiler toolchain does not guarantee ABI stability.
Hence, it does not make sense to build Qt with one version of compiler,
and then build the Qt application with other one.
Treat it as an error if there is mismatch between Emscripten version
used for building Qt and version used for building application.

Change-Id: Ia720da25ecbd048a009d0bf1927ba1bc1fa85fda
Reviewed-by: Morten Johan Sørvig <morten.sorvig@qt.io>
Reviewed-by: Ivan Tkachenko <me@ratijas.tk>