Long-live QT_REMOVAL_QT7_DEPRECATED_SINCE(major, minor)
... and QT_REMOVAL_QT8_DEPRECATED_SINCE.

Following up on the discussion from the Qt CS 2024 [0], this patch
introduces a new macro to mark the deprecated APIs that will be
removed in a specific major release. For now, add the macros
for Qt 7 and Qt 8.

The macro should be used instead of the regular
QT_DEPRECATED_SINCE(maj, min) for such type of APIs. The usage
is the same, as with the regular macro:

  #if QT_REMOVAL_QT7_DEPRECATED_SINCE(6, 9)
      QT_DEPRECATED_VERSION_X_6_9("The reason for the deprecation")
      void deprecatedFunc();
  #endif

The macro is basically a combination of QT_DEPRECATED_SINCE and
a comparison against a certain Qt major version.

[0]: https://wiki.qt.io/QtCS2024_Deprecate

Task-number: QTBUG-130024
Change-Id: I8efa1920d9d1524250bb00446aefffae65210340
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>