Fix an assertion failure in QLocale::uiLanguages()
Discovered by an Android 'phone configured with inconsistent system
locale and UI language list. The system locale specified more than the
en of what likely subtags expand to en-Latn-US and the single entry in
the UI language list was en-NO, so we fell back to adding the former,
but added the bcp47name() of the locale alongside the locale ID of the
CLDR entry for the system locale. The BCP47 name omits tags that what
remains will imply given likely-subtag rules, some of which may be
included in the CLDR entry's ID. This lead to an inconsistency
contradicting a later assertion.

Since we generally do want to search more specific names before less,
stick with the more specific ID from CLDR and use the name derived
from it as the uiLanguages entry to match, fixing the assertion. The
later post-processing of the list shall add the remaining
less-specific entries at suitable positions.

Change-Id: I1f1554dfb4c0198b04f8276fcb5d7869c9e58aac
Reviewed-by: Eimen Oueslati <eimen.oueslati@qt.io>
Reviewed-by: Nils Petter Skålerud <nils.petter.skalerud@qt.io>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>