macOS: Return high-DPI pixmaps from QCocoaTheme::standardPixmap
Most consumers feed the pixmaps into a QIcon, pulling out pixmaps
of sizes 32, 64, 128, 256, 512, so consumers of that QIcon will
not see low-resolution pixmaps for the common sizes like 64x64.

However if someone uses standardPixmap directly, or request an
icon with size 512x512, we won't have a high-DPI pixmap.

Unlike QStyle::standardPixmap we don't have a QWidget/QWindow
argument we can use to resolve the target DPR, so we have to
assume the pixmap can be used anywhere, and use the highest
DPR we've seen so far, via QGuiApplication::devicePixelRatio().

Change-Id: I3f41299b180f9361e3c7b1cd1f9bf4b1d00344da
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>