Windows: Return high-DPI pixmaps from QWindowsTheme::standardPixmap
Most consumers feed the pixmaps into a QIcon, pulling out pixmaps
of sizes 32, 64, 128, 256, 512, so consumers of that QIcon will
not see low-resolution pixmaps for the common sizes like 64x64.

However if someone uses standardPixmap directly, or request an
icon with size 512x512, we won't have a high-DPI pixmap.

Unlike QStyle::standardPixmap we don't have a QWidget/QWindow
argument we can use to resolve the target DPR, so we have to
assume the pixmap can be used anywhere, and use the highest
DPR we've seen so far, via QGuiApplication::devicePixelRatio().

Pick-to: 6.9
Change-Id: Ib7fb08ab4932da1ca3f3325e25a7022c1cd17435
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>