QBrush, QPen: Provide explicit comparison for types we compare with
Instead of implicitly creating a QPen or a QBrush from a color or style
just to be able to compare the result with the left-hand side pen or
brush, provide comparesEqual friends that can take a shortcut, avoiding
the allocation that comes with the implicit construction of the types.

operator==(Brush, BrushStyle) needs some extra handling, as we will
create a NoBrush brush if the style alone is not enough anyway.

Change-Id: I06e3d585160ad4fb5d12a46baf8c6a0aaaeeb0b0
Reviewed-by: Christian Ehrlicher <ch.ehrlicher@gmx.de>