tst_QLocale::defaulted_ctor(): actually use the default
The system locale isn't guaranteed to be the default; indeed, this is
one of the tests that frobs the default, so really shouldn't assume
it. Use the actual default and correct some related variable names in
the process.

This amends commit 3c8181de7006cab3c2d736dfb4adabe9b2ffaf4e.

Change-Id: I45dcdfb4f56295b0d22053937234347d33bbe320
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>