moc: StringRefStorage refactor without base classes
I had used base classes because that's what I had used for MetaTypeList,
where we did need to store a tuple-like collection of types.
Unfortunately, the side-effect of this is that we instantiate far more
types and this causes a huge increase in compilation times and, for GCC,
memory usage. That is especially an issue for qtopcua's
qopcuanodeids.cpp.

But it turns out we don't need them at all: instead of an collection of
base classes each of which contains a pointer, we can store an array of
pointers in our object. That reduces the number of types instantiated
from n to 1. This massively improves compilation times and memory usage:

           |       Previously             |         Now       |
Compiler   | Memory use | Time to compile | Memory    | Time  |
GCC 14.2   | 5066096 kB |            56 s | 673164 kB | 7.95s |
Clang 19.1 | ~594000 kB |            85 s | 481748 kB | 4.73s |

Drive-by change some variables to "int" to get overflow checking.

Change-Id: Id918c5597141c3a06ba4fffd58f9c09d9bba2401
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>