Fix restoration of default locale after tests that change it
Include an example of this in testlib's docs, as an illustration of
how to restore global state after a test that frobs it. Replace
various "restore as the test finishes" approaches with scope guard
forms that also work when the test fails. Add missing restoration to
tests that allowed their changes to leak into later tests. (If these
later tests actually depended on that, they should be amended to
control the default locale for themselves. This both makes the tests
more robust and documents the need for a specific locale.) Always
restore the actual prior default locale, rather than just assuming
it's the system locale (as several tests did).

For tst_QMimeDatabase this replaced a home-grown RAII class.
For tst_rcc, I included similar treatment of an unregistration of a
resource, since that clearly should follow the same pattern.

For tests that set the default locale in initTestCase(), I've moved
that to their constructor as that comes as close as possible to the
sense of doing it in main(), which is the natural place for a global
setting intended to apply to the whole program.

Change-Id: Ia7970a33b14c4fbd67c1d70dcafbba21f2fae3ef
Reviewed-by: Christian Ehrlicher <ch.ehrlicher@gmx.de>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>