tst_QMimeDatabase::installNewLocalMimeType() avoid spurious skip
In two of the four test contexts where this code is reused, the call
to comment(), another test function, from within this test function
skips (as it depends on FDO), which was leading to the "have we
failed" check after calling it to misleadingly return early. Instead,
duplicate the (complemented) skip condition in this call to avoid the
skip and let the rest of installNewLocalMimeType() run.

Change-Id: Icee039c0daa271e38b6657300503dd84219034f0
Reviewed-by: David Faure <david.faure@kdab.com>