QPen/QBrush: add optimized assignment operators for color and style
Assigning a color or style to an existing pen or brush (which we do
frequently in QPainter and elsewhere in Qt) used to implicitly create a
new QPen/QBrush instance and then move-assign that. This involves an
allocation of the new private data, and often also a deallocation of the
old pen or brush private if that wasn't shared (which it often isn't, as
we usually set a temporary object with no further copies).

By implementing specific assignment operators from color and style we
can shortcut this and avoid the allocations.

With this set of changes, the setPen/setBrush benchmarks now compare
to Qt 6.8 as follows:

setPen
6.8: 0.00016 msecs per iteration (total: 89, iterations: 524288)
6.9: 0.000032 msecs per iteration (total: 69, iterations: 2097152)

setBrush
6.8: 0.000078 msecs per iteration (total: 82, iterations: 1048576)
6.9: 0.000017 msecs per iteration (total: 73, iterations: 4194304)

i.e. improve by a factor of >4.5.

Change-Id: Ia9c0229b0f540ecf9afba2a598973c6931712f64
Reviewed-by: Christian Ehrlicher <ch.ehrlicher@gmx.de>