doc: Remove obsolete note after moving QFileSystemModel to QtGui
There doesn't seem to be any reason that it should require an instance
of QApplication. However, given that the model is by design
asynchronous, it does need to live in a thread that runs an event loop.
Add a note about that to the "Caching and Performance" section, and
apply some drive-by editing.

Amends fb9ec8ad44decba7b2878370f3711b61614f035a

Task-number: QTBUG-66177
Change-Id: Ibe3095e0264ce033732d1c724f59a871e2a0d3b8
Pick-to: 6.5 6.8
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>